refactor(cli): move update, await handlers
