+++ quartz/cli/bootstrap-cli.mjs
@@ -17,7 +17,7 @@ import serveHandler from "serve-handler"
 import { WebSocketServer } from "ws"
 import { randomUUID } from "crypto"
 import { Mutex } from "async-mutex"
-import { handleBuild } from "./handlers.js"
+import { handleBuild, handleCreate, handleUpdate } from "./handlers.js"
 import { CommonArgv, BuildArgv, CreateArgv, SyncArgv } from "./args.js"
 
 const ORIGIN_NAME = "origin"
@@ -65,24 +65,10 @@ yargs(hideBin(process.argv))
   .version(version)
   .usage("$0 <cmd> [args]")
   .command("create", "Initialize Quartz", CreateArgv, async (argv) => {
-    //
+    await handleCreate(argv)
   })
   .command("update", "Get the latest Quartz updates", CommonArgv, async (argv) => {
-    const contentFolder = path.join(cwd, argv.directory)
-    console.log(chalk.bgGreen.black(`\n Quartz v${version} \n`))
-    console.log("Backing up your content")
-    execSync(
-      `git remote show upstream || git remote add upstream https://github.com/jackyzha0/quartz.git`,
-    )
-    await stashContentFolder(contentFolder)
-    console.log(
-      "Pulling updates... you may need to resolve some `git` conflicts if you've made changes to components or plugins.",
-    )
-    gitPull(UPSTREAM_NAME, QUARTZ_SOURCE_BRANCH)
-    await popContentFolder(contentFolder)
-    console.log("Ensuring dependencies are up to date")
-    spawnSync("npm", ["i"], { stdio: "inherit" })
-    console.log(chalk.green("Done!"))
+    await handleUpdate(argv)
   })
   .command(
     "restore",
@@ -145,7 +131,7 @@ yargs(hideBin(process.argv))
     console.log(chalk.green("Done!"))
   })
   .command("build", "Build Quartz into a bundle of static HTML files", BuildArgv, async (argv) => {
-    handleBuild(argv)
+    await handleBuild(argv)
   })
   .showHelpOnFail(false)
   .help()
